Document Verification system/requirements.txt
streamlit
opencv-python
numpy
pillow
scikit-image
imutils
python-pptx
pymupdf
flask
